// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const ConnectionStatus = {
  PENDING: 'PENDING',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED',
  BLOCKED: 'BLOCKED'
} as const

export type ConnectionStatus = (typeof ConnectionStatus)[keyof typeof ConnectionStatus]


export const GroupVisibility = {
  PUBLIC: 'PUBLIC',
  PRIVATE: 'PRIVATE'
} as const

export type GroupVisibility = (typeof GroupVisibility)[keyof typeof GroupVisibility]


export const GroupMemberStatus = {
  PENDING: 'PENDING',
  ACTIVE: 'ACTIVE',
  REJECTED: 'REJECTED',
  BANNED: 'BANNED',
  LEFT: 'LEFT'
} as const

export type GroupMemberStatus = (typeof GroupMemberStatus)[keyof typeof GroupMemberStatus]


export const ReactionTargetType = {
  POST: 'POST',
  COMMENT: 'COMMENT'
} as const

export type ReactionTargetType = (typeof ReactionTargetType)[keyof typeof ReactionTargetType]


export const ReactionType = {
  LIKE: 'LIKE',
  LOVE: 'LOVE',
  HAHA: 'HAHA',
  SAD: 'SAD',
  ANGRY: 'ANGRY'
} as const

export type ReactionType = (typeof ReactionType)[keyof typeof ReactionType]


export const StorageProvider = {
  SUPABASE: 'SUPABASE'
} as const

export type StorageProvider = (typeof StorageProvider)[keyof typeof StorageProvider]


export const ChatMemberRole = {
  OWNER: 'OWNER',
  MEMBER: 'MEMBER'
} as const

export type ChatMemberRole = (typeof ChatMemberRole)[keyof typeof ChatMemberRole]


export const ChatMessageType = {
  TEXT: 'TEXT',
  MEDIA: 'MEDIA',
  SYSTEM: 'SYSTEM'
} as const

export type ChatMessageType = (typeof ChatMessageType)[keyof typeof ChatMessageType]